work on PR comments
